pyproject.toml
[project]
name = "sora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.4"
dependencies = [
    "openai>=2.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "pillow>=10.4.0",
]
